+++ src/main.js
@@ -1,10 +1,10 @@
-const { setupProgram } = require('./program')
-const { startDevServer: startDevServer } = require('./dev-server')
+const { createProgram } = require('./program')
+const { startDevServer } = require('./dev-server')
 /**
  * The main entry point for the CLI. Creates a Commander program and starts the Next.js Custom Dev Server
  */
 function main() {
-  const program = setupProgram()
+  const program = createProgram()
 
   startDevServer(program)
 }
