+++ src/main.test.js
@@ -1,20 +1,20 @@
 const { main } = require('./main')
-const { startDevServer: startDevServer } = require('./dev-server')
-const { setupProgram } = require('./program')
+const { startDevServer } = require('./dev-server')
+const { createProgram } = require('./program')
 
 jest.mock('./dev-server', () => ({ startDevServer: jest.fn() }))
-jest.mock('./program', () => ({ setupProgram: jest.fn() }))
+jest.mock('./program', () => ({ createProgram: jest.fn() }))
 
 const mockProgram = jest.fn()
 
 describe('main', () => {
   beforeEach(jest.clearAllMocks)
   test('should setup the program and start the app', () => {
-    setupProgram.mockReturnValueOnce(mockProgram)
+    createProgram.mockReturnValueOnce(mockProgram)
 
     main()
 
-    expect(setupProgram).toHaveBeenCalledWith()
+    expect(createProgram).toHaveBeenCalledWith()
     expect(startDevServer).toHaveBeenCalledWith(mockProgram)
   })
 })
