+++ src/program/create.js
@@ -6,7 +6,7 @@ const { defaultWatchEvent } = require('../constants')
  * This function passes the different configuration options to the Commander program and return the program itself.
  * @returns Commander program
  */
-function setupProgram() {
+function create() {
   program.version(pkg.version)
   program
     .option('-r, --root [dir]', 'root directory of your nextjs app')
@@ -30,4 +30,4 @@ function setupProgram() {
   return program
 }
 
-module.exports = setupProgram
+module.exports = create
